Move out options from engines util & create skeleton for regional filters
