+++ extension-manifest-v3/src/background/index.js
@@ -22,9 +22,13 @@ import './serp.js';
 import './helpers.js';
 import './external.js';
 
+import './filters/custom-filters.js';
+import './filters/experimental-filters.js';
+import './filters/regional-filters.js';
+
+import './exceptions.js';
+
 import './telemetry.js';
 import './reporting/index.js';
 
 import './devtools.js';
-import './custom-filters.js';
-import './exceptions.js';
